requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
bcrypt==5.0.0
PyYAML==6.0.3
PyJWT==2.10.1
supabase==2.27.0